Loading and Cleaning Data
Created a function to load all reddit posts and create a dictionary of lists of each post, where the dictionary will then be used to create a feature and label set.

The next step will be creating two lists: one of the features and one of the labels using the dictionary.