Generative Adversarial Network for Galaxy Zoo
A deep generative adversarial network using the galaxy zoo database. Not only are adversarial attacks on image classifiers important in understanding pitfalls of a regular convolutional neural network, but it can also be used to counteract overfitting in image classifying models. Thanks to the game theory of it, the adversarial attacks train the convolutional network to be more robust.